fix: Address review feedback for channel storage refactor [Midtown !1646]
- Move tests to separate channel_tests.rs per CLAUDE.md convention
- Fix migration orphan bug: append content when both V1 and V2 exist
- Add migration guarantee comment to cursor.rs file_path()
- Document channel storage layout in docs/architecture.md
- Add regression test for orphan migration scenario
